requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
black==20.8b1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dataclasses==0.7
defusedxml==0.7.0rc1
diff-match-patch==20200713
Django==3.0.1
django-admin-autocomplete-list-filter==0.1.3
django-braces==1.14.0
django-etc==1.2.0
django-excel==0.0.10
django-filter==2.0.0
django-import-export==2.4.0
django-model-utils==4.0.0
django-oauth-toolkit==1.4.1
django-rest-registration==0.5.6
django-sendgrid-v5==0.9.0
django-star-ratings==0.9.1
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
drf-social-oauth2==1.0.9
et-xmlfile==1.0.1
future==0.18.2
idna==2.10
isort==5.4.2
jdcal==1.4.1
lazy-object-proxy==1.4.3
lml==0.1.0
MarkupPy==1.14
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pathspec==0.8.0
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
pycparser==2.20
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
PyJWT==2.0.1
pylint==2.6.0
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==6.4.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
swapper==1.1.2.post1
tablib==3.0.0
texttable==1.6.3
toml==0.10.1
typed-ast==1.4.1
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.25.9
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
